requirements.txt
Flask==3.1.1
python-dotenv==1.0.1
openai==1.43.0
twilio==9.2.3
gunicorn==22.0.0
